// clients/securityService/models/SecurityRoleCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SecurityRoleCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name: string;
    tenantId: string;
    description?: string | null;
};